Updated remaining components to add userId to each item sent to the database so that only information created by a user is retrieved. Added authorization guards to all protected urls to prevent unauthorized access to user information.

+++ bronco-resume-app/backend/routes/projects.js
@@ -1,15 +1,17 @@
 const express = require('express');
-
+const checkAuth = require('../middleware/check-auth');
 const router = express.Router();
 
 const Project = require('../models/project');
 
-router.post('', (req, res, next) => {
+router.post('',
+  checkAuth, (req, res, next) => {
   const project = new Project({
     title: req.body.title,
     startDate: req.body.startDate,
     endDate: req.body.endDate,
-    description: req.body.description
+    description: req.body.description,
+    creator: req.userData.userId
   });
   project.save().then(createdProject => {
     res.status(201).json({
@@ -19,8 +21,9 @@ router.post('', (req, res, next) => {
   });
 });
 
-router.get('', (req, res, next) => {
-  Project.find()
+router.get('',
+  checkAuth, (req, res, next) => {
+  Project.find({creator: req.userData.userId})
   .then(documents => {
     res.status(200).json({
       message: 'Projects fetched successfully!',
@@ -29,9 +32,41 @@ router.get('', (req, res, next) => {
   })
 });
 
-router.delete('/:id', (req, res, next) => {
-  Project.deleteOne({ _id: req.params.id }).then(result => {
-   console.log(result);
+router.get('/:id',
+  checkAuth, (req, res, next) => {
+  Project.findById(req.params.id).then(project => {
+    if (project) {
+      res.status(200).json(project);
+    } else {
+      res.status(404).json({message: 'Award not found!'});
+    }
+  })
+  .then(documents => {
+    res.status(200).json({
+      message: 'Project fetched successfully!',
+      project: documents
+    })
+  })
+});
+
+router.put('/:id',
+  checkAuth, (req, res, next) => {
+  const project = new Project({
+    title: req.body.title,
+    startDate: req.body.startDate,
+    endDate: req.body.endDate,
+    description: req.body.description,
+    _id: req.body.id
+  });
+  Project.updateOne({_id: req.params.id, creator: req.userData.userId}, project)
+    .then(result => {
+      res.status(200).json({message: "Project update successful!"})
+    })
+});
+
+router.delete('/:id',
+  checkAuth, (req, res, next) => {
+  Project.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
    res.status(200).json({ message: "Project deleted!" });
   });
 });
